fix some comments from morning call
