+++ src/components/Buttons/ResetZoomButton.tsx
