+++ src/pages/landing-page/index.tsx
@@ -1,7 +1,7 @@
 // import React from 'react'
 import jenkinsButler from '../../assets/jenkins-butler.svg'
-import NavBar from '../../components/NavBar'
-import Footer from '../../components/Footer'
+import NavBar from '../../components/Layout/NavBar'
+import Footer from '../../components/Layout/Footer'
 import './landing-page.css'
 import { Link, NavLink } from 'react-router-dom'
 import { styled } from '@mui/system'
