+++ src/pages/plugin-detail/index.tsx
@@ -4,11 +4,11 @@ import { Stack, Typography, Card, Grid, Link, Button, Box } from '@mui/material'
 import { styled } from '@mui/system'
 
 // import Grid from '@mui/material/Unstable_Grid2'
-import NavBar from '../../components/NavBar'
-import PluginInstallationsChart from '../../components/PluginInstallationsChart'
-import PluginInstallationsPerVersion from '../../components/PluginInstallationsPerVersion'
-import PluginInstallationsPercentageChart from '../../components/PluginInstallationsPercentageChart'
-import PluginInstallationsPercentagePerVersionChart from '../../components/PluginInstallationsPercentagePerVersionChart'
+import NavBar from '../../components/Layout/NavBar'
+import PluginInstallationsChart from '../../components/Charts/PluginInstallationsChart'
+import PluginInstallationsPerVersion from '../../components/Charts/PluginInstallationsPerVersionChart'
+import PluginInstallationsPercentageChart from '../../components/Charts/PluginInstallationsPercentageChart'
+import PluginInstallationsPercentagePerVersionChart from '../../components/Charts/PluginInstallationsPercentagePerVersionPieChart'
 
 import InsertDriveFileIcon from '@mui/icons-material/InsertDriveFile'
 import ImageIcon from '@mui/icons-material/Image'
