+++ src/pages/plugin-trends/index.tsx
@@ -11,11 +11,11 @@ import {
     Select,
     MenuItem,
 } from '@mui/material'
-import NavBar from '../../components/NavBar'
+import NavBar from '../../components/Layout/NavBar'
 import useFetchAndFilterPlugins from '../../hooks/useFetchAndFilterPlugins'
 import useSortPlugins from '../../hooks/useSortPlugins'
 import usePagination from '../../hooks/usePagination'
-import PluginCard from '../../components/PluginCard'
+import PluginCard from '../../components/Plugins/PluginCard'
 
 const PluginTrends: React.FC = () => {
     const [searchTerm, setSearchTerm] = useState<string>('')
